src/quantum_blockchain.py
import asyncio
import hashlib
import json
import random
import time
from datetime import datetime
from typing import Dict, Any

import boto3
from braket.aws import AwsDevice
from braket.circuits import Circuit
from braket.devices import LocalSimulator


class QuantumService:
    """Service for quantum operations using AWS Braket"""

    def __init__(self):
        self.local_simulator = LocalSimulator()
        # For production, you would use: AwsDevice("arn:aws:braket:::device/quantum-simulator/amazon/sv1")

    def generate_quantum_random_number(self, seed_text: str) -> int:
        """Generate a quantum random number based on user input"""
        # Create a quantum circuit for random number generation
        circuit = Circuit()

        # Use the seed text to determine number of qubits (between 3 and 8)
        num_qubits = 3 + (hash(seed_text) % 6)

        # Create superposition states
        for i in range(num_qubits):
            circuit.h(i)

        # Add some rotation based on the seed
        for i, char in enumerate(seed_text[:num_qubits]):
            angle = (ord(char) / 128.0) * 3.14159  # Convert char to angle
            circuit.rx(i, angle)

        # Measure all qubits
        for i in range(num_qubits):
            circuit.measure(i)

        try:
            # Run on local simulator for demo (replace with AWS Braket in production)
            task = self.local_simulator.run(circuit, shots=10)
            result = task.result()
            measurement = result.measurement_counts

            # Convert binary result to integer
            binary_result = list(measurement.keys())[0]
            quantum_number = int(binary_result, 2)

            return quantum_number
        except Exception as e:
            print(f"Quantum simulation error: {e}")
            # Fallback to quantum-inspired random number
            return hash(seed_text + str(time.time())) % 1000

    def create_bell_state_circuit(self) -> Dict[str, float]:
        """Create a Bell state and measure correlations"""
        circuit = Circuit()

        # Create Bell state |00⟩ + |11⟩
        circuit.h(0)  # Put first qubit in superposition
        circuit.cnot(0, 1)  # Entangle with second qubit

        # Measure both qubits
        circuit.measure(0)
        circuit.measure(1)

        try:
            # Run multiple shots to get probability distribution
            task = self.local_simulator.run(circuit, shots=1000)
            result = task.result()
            counts = result.measurement_counts

            # Normalize to probabilities
            total_shots = sum(counts.values())
            probabilities = {}

            for state in ['00', '01', '10', '11']:
                probabilities[state] = counts.get(state, 0) / total_shots

            return probabilities
        except Exception as e:
            print(f"Bell state simulation error: {e}")
            # Return theoretical Bell state probabilities
            return {'00': 0.5, '01': 0.0, '10': 0.0, '11': 0.5}


class BlockchainService:
    """Service for blockchain operations (simulated QLDB functionality)"""

    def __init__(self):
        # In production, you would initialize AWS QLDB client
        # self.qldb_client = boto3.client('qldb')
        pass

    def store_quantum_result(self, name: str, message: str, quantum_number: int) -> Dict[str, Any]:
        """Store quantum result in blockchain ledger"""
        # Create transaction data
        transaction_data = {
            'name': name,
            'message': message,
            'quantum_number': quantum_number,
            'timestamp': datetime.now().isoformat(),
            'type': 'quantum_generation'
        }

        # Generate transaction ID
        transaction_id = self._generate_transaction_id(transaction_data)

        # Generate block hash (simulated)
        block_hash = self._generate_block_hash(transaction_data, transaction_id)

        # In production, you would insert into QLDB:
        # self._insert_to_qldb(transaction_data, transaction_id, block_hash)

        return {
            'transaction_id': transaction_id,
            'block_hash': block_hash,
            'timestamp': transaction_data['timestamp']
        }

    def _generate_transaction_id(self, data: Dict[str, Any]) -> str:
        """Generate unique transaction ID"""
        data_string = json.dumps(data, sort_keys=True)
        return hashlib.sha256(data_string.encode()).hexdigest()[:16]

    def _generate_block_hash(self, data: Dict[str, Any], transaction_id: str) -> str:
        """Generate block hash"""
        combined = f"{transaction_id}{json.dumps(data, sort_keys=True)}{time.time()}"
        return hashlib.sha256(combined.encode()).hexdigest()

    def _insert_to_qldb(self, transaction_data: Dict[str, Any], transaction_id: str, block_hash: str):
        """Insert data to AWS QLDB (placeholder for production implementation)"""
        # Production implementation would use QLDB driver:
        # with create_qldb_driver() as driver:
        #     driver.execute_lambda(lambda executor:
        #         executor.execute_statement("INSERT INTO QuantumTransactions ?", transaction_data))
        pass


class DemoQuantumBlockchain:
    """Main service combining quantum and blockchain functionality"""

    def __init__(self):
        self.quantum_service = QuantumService()
        self.blockchain_service = BlockchainService()

    async def process_user_input(self, name: str, message: str) -> Dict[str, Any]:
        """Process user input and generate quantum + blockchain results"""
        try:
            # Generate quantum random number
            quantum_number = self.quantum_service.generate_quantum_random_number(f"{name}:{message}")

            # Create entanglement demonstration
            entanglement_data = self.quantum_service.create_bell_state_circuit()

            # Store in blockchain
            blockchain_result = self.blockchain_service.store_quantum_result(name, message, quantum_number)

            return {
                'success': True,
                'quantum_number': quantum_number,
                'entanglement_data': [
                    entanglement_data['00'],
                    entanglement_data['01'],
                    entanglement_data['10'],
                    entanglement_data['11']
                ],
                'transaction_id': blockchain_result['transaction_id'],
                'block_hash': blockchain_result['block_hash'],
                'timestamp': blockchain_result['timestamp']
            }
        except Exception as e:
            return {
                'success': False,
                'error': str(e)
            }