src/pyproject.toml
[project]
name = "qc-bc-interactive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiohttp>=3.13.0",
    "amazon-braket-sdk>=1.102.6",
    "boto3>=1.40.50",
    "debugpy>=1.8.17",
    "fastapi>=0.119.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
